@@ -8,12 +8,12 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "axios": "^1.6.8",
     "firebase": "^10.11.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-image-lightbox": "^5.1.1",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.3.4",
-    "react-s3": "^1.0.7",
+    "react-router": "^6.23.1",
+    "react-router-dom": "^6.23.1",
+    "react-s3": "^1.3.1",
     "react-scripts": "^5.0.1",
     "react-uuid": "^1.0.2"
   },
@@ -39,6 +39,6 @@
     ]
   },
   "devDependencies": {
-    "dotenv": "^8.2.0"
+    "dotenv": "^16.4.5"
   }
 }
