@@ -1,26 +1,30 @@
 import React, { Component } from "react";
-import { Route, Switch, Redirect } from "react-router-dom";
 import Signup from "./Signup";
 import Login from "./Login";
 import Profile from "./Profile";
 import Feed from "./Feed";
-import Dummy from "./Dummy";
 import OTPInput from "./OTPInput";
+
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+
 
 
 class Main extends Component {
   render() {
-    return (
-      <Switch>
-        <Route path="/" exact component={Login} />
-        <Route path="/signup" exact component={Signup} />
-        <Route path="/otpinput" exact component={OTPInput} />
-        <Route path="/login" exact component={Login} />
-        <Route path="/profile/:id?" exact component={Profile} />
-        <Route path="/feed" exact component={Feed} />
-        <Redirect to="/not-found" />
-      </Switch>
+      return (
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Login />} />
+          <Route path="/signup" element={<Signup />} />
+          <Route path="/otpinput" element={<OTPInput />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/profile/:id" element={<Profile />} />
+          <Route path="/feed" element={<Feed />} />
+          {/* <Route path="*" element={<NotFound />} /> */}
+        </Routes>
+      </BrowserRouter>
     );
+   
   }
 }
 
